+++ src/TurtlePath/Commands/UpdateCommandHandler.cs
@@ -1,19 +1,11 @@
 namespace TurtlePath.Commands
 {
-    using Microsoft.Extensions.DependencyInjection;
-    using TurtlePath.Exceptions;
-    using TurtlePath.Hooks;
-    using TurtlePath.Models.Requests;
-    using TurtlePath.Models.Responses;
-    using TurtlePath.Persistence;
-    using TurtlePath.Validation;
-    using TurtlePath.Mapping;
-    using TurtlePath.Domain.Contracts;
-    using TurtlePath.Domain.Identifier;
     using Pelican.Mediator;
     using System;
-    using System.Threading;
-    using System.Threading.Tasks;
+    using TurtlePath.Domain.Contracts;
+    using TurtlePath.Domain.Identifier;
+    using TurtlePath.Models.Requests;
+    using TurtlePath.Models.Responses;
 
     /// <summary>
     /// Provides a base implementation for handling update commands for TurtlePath BaseEntity instances with CId identifiers.
@@ -40,7 +32,7 @@ namespace TurtlePath.Commands
     /// </summary>
     /// <typeparam name="TRequest">The type of the request.</typeparam>
     /// <typeparam name="TEntity">The type of the entity being updated.</typeparam>
-    public abstract class UpdateCommandHandler<TRequest, TEntity> : GenericUpdateNoReturnCommandHandler<TRequest, TEntity, CId>
+    public abstract class UpdateCommandHandler<TRequest, TEntity> : GenericUpdateCommandHandler<TRequest, TEntity, CId>
         where TRequest : BaseRequest, IRequest
         where TEntity : BaseEntity
     {